.claude/skills/tech-artist-generate-asset/requirements.txt
# Python dependencies for tech-artist-generate-asset skill
# Install with: pip install -r requirements.txt

# Image processing library
Pillow>=10.0.0

# Numerical computing (for image operations)
numpy>=1.24.0
